- Added webpack - Renamed files appropriately - Cleaned up code
