+++ package.json
@@ -5,8 +5,8 @@
   "main": "index.js",
   "scripts": {
     "start": "node ./server/server.js",
-    "dev": "nodemon ./server/server.js -w ./server",
     "test": "karma start",
+    "build": "webpack --watch",
     "tsc": "tsc -p . -w --outDir build"
   },
   "keywords": [],
@@ -29,6 +29,7 @@
     "karma-chrome-launcher": "^0.2.3",
     "karma-jasmine": "^0.3.8",
     "karma-ng-html2js-preprocessor": "^0.2.2",
+    "ts-loader": "^0.9.5",
     "typescript": "^2.0.3",
     "typings": "^1.4.0"
   }
