--- public/admin/userList.ts
@@ -1,20 +1,0 @@
-angular.module('app').component('userList', {
-  templateUrl: 'admin/userList.html',
-  bindings: {
-    allUsers: '='
-  },
-  controller: class userListCtrl {
-    allUsers: any;
-    users: any;
-
-    constructor() {
-      this.allUsers.sort((user1: any, user2: any) => {
-        if(user1.firstName < user2.firstName) return -1;
-        if(user1.firstName === user2.firstName) return 0;
-        if(user1.firstName > user2.firstName) return 1;
-      })
-
-      this.users = this.allUsers;
-    }
-  }
-})
