+++ public/components/talkDuration.filter.ts
@@ -1,0 +1,4 @@
+export default angular.module('app.talkDuration', [])
+    .filter('talkDuration', () => {
+        return (duration: string) => `Duration: ${duration} minutes`;
+    })
