--- public/components/talkDuration.ts
@@ -1,3 +1,0 @@
-angular.module('app').filter('talkDuration', () => {
-    return (duration: string) => `Duration: ${duration} minutes`;
-})
