+++ public/index.html
@@ -6,41 +6,27 @@
   <link rel="stylesheet" href="//cdnjs.cloudflare.com/ajax/libs/toastr.js/latest/css/toastr.min.css">
   <link rel="stylesheet" href="/styles.css">
   <link rel="shortcut icon" type="image/x-icon" href="/favicon.ico">
+
+  <!-- 3rd party -->
+  <script src="/vendor/jquery.min.js"></script>
+  <script src="/vendor/1.5.5/angular.min.js"></script>
+  <script src="/vendor/1.5.5/angular-route.min.js"></script>
+  <script src="/vendor/toastr.min.js"></script>
+  <script>
+    (function(orig) {
+      angular.modules = [];
+      angular.module = function() {
+        if (arguments.length > 1) {
+          angular.modules.push(arguments[0]);
+        }
+        return orig.apply(null, arguments);
+      }
+    })(angular.module);
+  </script>
+
 </head>
 <body class="container">
   <ng-view></ng-view>
-
-  <script type="text/javascript" src="/vendor/jquery.min.js"></script>
-  <script type="text/javascript" src="vendor/1.5.5/angular.min.js"></script>
-  <script type="text/javascript" src="vendor/1.5.5/angular-route.min.js"></script>
-  <script type="text/javascript" src="vendor/toastr.min.js"></script>
-  <script type="text/javascript" src="/toastr/toastr.js"></script>
-  <script type="text/javascript" src="/app.js"></script>
-  <script type="text/javascript" src="/routes.js"></script>
-  <script type="text/javascript" src="/admin/adminLogin.js"></script>
-  <script type="text/javascript" src="/admin/results.js"></script>
-  <script type="text/javascript" src="/admin/createUsers.js"></script>
-  <script type="text/javascript" src="/admin/userList.js"></script>
-  <script type="text/javascript" src="/admin/userDetails.js"></script>
-  <script type="text/javascript" src="/admin/nameParser.js"></script>
-  <script type="text/javascript" src="/nav/nav.js"></script>
-  <script type="text/javascript" src="/security/logout.js"></script>
-  <script type="text/javascript" src="/security/login.js"></script>
-  <script type="text/javascript" src="/security/auth.js"></script>
-  <script type="text/javascript" src="/security/currentIdentity.js"></script>
-  <script type="text/javascript" src="/security/users.js"></script>
-  <script type="text/javascript" src="/home/home.js"></script>
-  <script type="text/javascript" src="/home/createNewSession.js"></script>
-  <script type="text/javascript" src="/home/unreviewedTalk.js"></script>
-  <script type="text/javascript" src="/sessions/sessions.js"></script>
-  <script type="text/javascript" src="/sessions/unreviewedSessionCount.js"></script>
-  <script type="text/javascript" src="/sessions/sessionDetail.js"></script>
-  <script type="text/javascript" src="/sessions/sessionDetailWithVotes.js"></script>
-  <script type="text/javascript" src="/components/detailPanel.js"></script>
-  <script type="text/javascript" src="/components/talkDuration.js"></script>
-  <script type="text/javascript" src="/components/zoomin.js"></script>
-  <script type="text/javascript" src="/profile/profile.js"></script>
-  
-
+  <script src="bundle.js"></script>
 </body>
 </html>
