+++ public/sessions/sessionDetail.component.ts
@@ -1,0 +1,13 @@
+class sessionDetailCtrl {
+  constructor() {}
+}
+
+export default angular.module('app.sessionDetail', [])
+  .component('sessionDetail', {
+    templateUrl: '/sessions/sessionDetail.html',
+    bindings: {
+    session: '=',
+      initialCollapsed: '@'
+    },
+    controller: sessionDetailCtrl
+  })
