--- public/sessions/sessionDetail.ts
@@ -1,10 +1,0 @@
-angular.module('app').component('sessionDetail', {
-  templateUrl: '/sessions/sessionDetail.html',
-  bindings: {
-    session: '=',
-    initialCollapsed: '@'
-  },
-  controller: class sessionDetailCtrl {
-    constructor() {}
-  }
-})
