--- public/sessions/sessionDetailWithVotes.ts
@@ -1,10 +1,0 @@
-angular.module('app').component('sessionDetailWithVotes', {
-  templateUrl: '/sessions/sessionDetailWithVotes.html',
-  bindings: {
-    session: '=',
-    initialCollapsed: '@'
-  },
-  controller: class sessionDetailCtrl {
-    constructor() {}
-  }
-})
