+++ public/toastr/toastr.ts
@@ -1,12 +1,8 @@
 // Tell TypeScript that toastr has already been declared (Global variable)
 declare let toastr: any;
+toastr.options.timeOut = 1000;
+export default angular.module('toastr', []).value('toastr', toastr);
 
-(function() {
 
-  let toastrModule = angular.module('toastr', []);
 
-  toastr.options.timeOut = 1000;
 
-  toastrModule.value('toastr', toastr);
-
-}())
